SONATA actor try to reconnect to Kafka brokers rather than failing
SONATA actor was failing on error, now it keep trying forever to reconnect
to Kafka brokers. The delay between reconnect events is 60 seconds for
all brokers down event, and 30 seconds for all other errors.
SONATA still fails when the SONATA topic doesn't exist, this is okay
because it is checked on startup.

The number of retries is counted in opentelemetry under
`netgauze.collector.flows.sonata.kafka.reconnect.attempts`.